避免重复安装导致共享路径覆盖、matlabrc重复操作的问题

+++ 提权操作/Program.vb
@@ -48,8 +48,13 @@ Module Program
 				Directory.CreateDirectory(路径)
 				路径 = Path.Combine(路径, "共享路径.txt")
 				'必须用文本文件记录路径，系统环境变量会被用户环境变量屏蔽
-				File.Create(路径)
-				File.AppendAllLines(Path.Combine(MatlabRoot, "toolbox\local\matlabrc.m"), {$"path(path,fileread('{路径}'));%PathManager"})
+				If Not File.Exists(路径) Then
+					'不要覆盖已有文件
+					File.Create(路径)
+				End If
+				路径 = Path.Combine(MatlabRoot, "toolbox\local\matlabrc.m")
+				'此处要注意避免重复安装导致重复添加多行
+				File.WriteAllLines(路径, (From 行 As String In File.ReadAllLines(路径) Select 行 Where Not 行.EndsWith("%PathManager")).Append($"path(path,fileread('{路径}'));%PathManager"))
 			Case 提权操作.Uninstall_Path_Manager
 				Dim Matlab路径 As String = 读入字符串(参数流)
 				File.Copy(Path.Combine(Path.GetDirectoryName(Assembly.GetEntryAssembly.Location), "..\savepath.m"), Path.Combine(Matlab路径, "toolbox\matlab\general\savepath.m"), True)
